refactor: LadderPrinter 패키지 이동
