+++ src/main/java/view/RowPrinter.java
@@ -1,12 +1,11 @@
-package domain;
+package view;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 class RowPrinter {
-    static String from(Row row) {
-        List<Boolean> rowInfos = row.getBridges();
-        String rawRowString = rowInfos.stream().map(RowPrinter::makeBridge)
+    static String from(List<Boolean> bridges) {
+        String rawRowString = bridges.stream().map(RowPrinter::makeBridge)
                 .collect(Collectors.joining("|"));
         return "    |%s|".formatted(rawRowString);
     }
